Report failure when ~error-on-warn (fix #6)
When compiling with `byte-compile-file`, warnings are reported as
errors but this won't stop compilation. With this patch, any error or
warning will exit Emacs with non-zero return value.

This commit also changes the error message in case of
problems. Instead of repeating the first error message with less
information (i.e., the filename, line and column numbers disappear),
the message "There were compile-time errors" appear instead.