Only download each package archive once per package-user-dir
This should cut down the number of times we pull each package archive
to a max of twice per build: once for the package and once for any
dependencies of the tests.

It may be possible to get this down to exactly once, but that will
take more doing and can work alongside this solution.

Related: #11