Let PACKAGE_FILE be the *-pkg.el file
In some projects (e.g., Magit, Forge), the *-pkg.el file is in a
directory above the actual lisp, so PACKAGE-pkg.el will not be found
next to PACKAGE.el.

The only real thing this impacts outside determining the
package-descriptor is automatic handling of `load-path', which is
extensible via Makefile.