Introduce EMAKE_USE_LOCAL and deprecate PACKAGE_IGNORE_DEPS
Instead of relying on the maintainer to manage the load-path to
include packages in parallel development, this enhancement attempts to
do so automatically.  The biggest difficulty in development solved by
these new enhancements is the problem where

  - a package dependency chain A > B > C exists,

  - a development version of (B) is loaded from disk, and

  - (C) must still be installed.

So, the dependencies of (B) must be proactively chased-down, which
also means making some assumptions about the file that provides this
dependency information (e.g., `PACKAGE-pkg.el`).  This assumption can
be 'fixed' where need be with `emake-package-dev-locations-functions`.